Refactor import statements and format code
The import statements in the Settings layout and related files have been cleaned up and the code has been formatted for better readability. The import path for SettingsSidenav has been shortened, removing the repetition. Extra whitespaces and lines were also removed to adhere to the coding standards.
