+++ packages/apps/spaces/app/settings/page.tsx
@@ -1,22 +1,23 @@
 import React from 'react';
 
-import {TabsContainer} from "./Tabs/TabsContainer";
-import {Panels} from "./Tabs/Panels";
-import {SettingsMainSection} from "./SettingsMainSection";
+import { TabsContainer } from './Tabs/TabsContainer';
+import { Panels } from './Tabs/Panels';
+import { SettingsMainSection } from './SettingsMainSection';
 
 interface SettingsPageProps {
-    searchParams: { tab?: string };
+  searchParams: { tab?: string };
 }
 
-export default async function SettingsPage({searchParams}: SettingsPageProps) {
-
-    return (
-        <>
-            <SettingsMainSection>
-                <TabsContainer>
-                    <Panels tab={searchParams.tab ?? 'oauth'} />
-                </TabsContainer>
-            </SettingsMainSection>
-        </>
-    );
-};
+export default async function SettingsPage({
+  searchParams,
+}: SettingsPageProps) {
+  return (
+    <>
+      <SettingsMainSection>
+        <TabsContainer>
+          <Panels tab={searchParams.tab ?? 'oauth'} />
+        </TabsContainer>
+      </SettingsMainSection>
+    </>
+  );
+}
